add swagger docs
