+++ model/team.go
@@ -16,7 +16,7 @@ package model
 
 // Team represents a team or organization in the remote version control system.
 //
-// swagger:model user
+// swagger:model team
 type Team struct {
 	// Login is the username for this team.
 	Login string `json:"login"`
